Command line converter and GUI converter are separate applications.
